requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
et-xmlfile==1.1.0
fake-useragent==1.1.1
frozenlist==1.3.3
idna==3.4
importlib-resources==5.10.2
lxml==4.9.2
multidict==6.0.4
openpyxl==3.0.10
Pillow==9.4.0
requests==2.28.1
soupsieve==2.3.2.post1
tqdm==4.64.1
urllib3==1.26.13
yarl==1.8.2
zipp==3.11.0